refactor: Extract config consts from locales file since lingui needs it and messages are not compiled yet
