+++ src/pages/sunshine/[entity]/[id]/power-stability.tsx
@@ -31,7 +31,7 @@ import {
   SunshinePowerStabilityData,
 } from "src/domain/data";
 import { getLocalizedLabel } from "src/domain/translation";
-import { defaultLocale } from "src/locales/locales";
+import { defaultLocale } from "src/locales/config";
 
 type Props =
   | (Extract<SharedPageProps, { entity: "operator"; status: "found" }> & {
